+++ DfE.FindInformationAcademiesTrusts/Pages/ManageProjectsAndCases/Overview/ResultSorting.cs
@@ -1,0 +1,10 @@
+﻿namespace DfE.FindInformationAcademiesTrusts.Pages.ManageProjectsAndCases.Overview
+{
+    public static class ResultSorting
+    {
+        public const string CreatedDesc = "createdDesc";
+        public const string CreatedAsc = "createdAsc";
+        public const string UpdatedDesc = "updatedDesc";
+        public const string UpdatedAsc = "updatedAsc";
+    }
+}